Reply to latest whisper command
