+++ src/Sidekick.Business/StartupExtensions.cs
@@ -16,7 +16,6 @@ using Sidekick.Business.Http;
 using Sidekick.Business.ItemCategories;
 using Sidekick.Business.Languages;
 using Sidekick.Business.Parties;
-using Sidekick.Business.Whispers;
 using Sidekick.Business.Windows;
 using Sidekick.Domain.Languages;
 
@@ -36,7 +35,6 @@ namespace Sidekick.Business
             services.AddSingleton<IPoePriceInfoClient, PoePriceInfoClient>();
             services.AddSingleton<ITradeSearchService, TradeSearchService>();
             services.AddSingleton<IWindowService, WindowService>();
-            services.AddSingleton<IWhisperService, WhisperService>();
 
             services.AddSingleton<IPoeDbClient, PoeDbClient>();
             services.AddSingleton<IPoeWikiClient, PoeWikiClient>();
