--- src/Sidekick.Business/Whispers/IWhisperService.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sidekick.Business.Whispers
-{
-    public interface IWhisperService
-    {
-        Task<bool> ReplyToLatestWhisper();
-    }
-}
