+++ src/Sidekick.Domain/Game/Chat/Commands/ReplyToLatestWhisperCommand.cs
@@ -1,0 +1,11 @@
+using MediatR;
+
+namespace Sidekick.Domain.Game.Chat.Commands
+{
+    /// <summary>
+    /// Start writing to the last person that has messaged the player inside Path of Exile
+    /// </summary>
+    public class ReplyToLatestWhisperCommand : ICommand<bool>
+    {
+    }
+}
