+++ src/Sidekick.Infrastructure/Game/Chat/GoToHideoutHandler.cs
@@ -4,14 +4,14 @@ using MediatR;
 using Sidekick.Domain.Game.Chat.Commands;
 using Sidekick.Domain.Process;
 
-namespace Sidekick.Application.Game.Chat
+namespace Sidekick.Infrastructure.Game.Chat
 {
-    public class ExitToCharacterSelectionHandler : ICommandHandler<ExitToCharacterSelectionCommand, bool>
+    public class GoToHideoutHandler : ICommandHandler<GoToHideoutCommand, bool>
     {
         private readonly IMediator mediator;
         private readonly INativeProcess nativeProcess;
 
-        public ExitToCharacterSelectionHandler(
+        public GoToHideoutHandler(
             IMediator mediator,
             INativeProcess nativeProcess)
         {
@@ -19,14 +19,14 @@ namespace Sidekick.Application.Game.Chat
             this.nativeProcess = nativeProcess;
         }
 
-        public async Task<bool> Handle(ExitToCharacterSelectionCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(GoToHideoutCommand request, CancellationToken cancellationToken)
         {
             if (!nativeProcess.IsPathOfExileInFocus)
             {
                 return false;
             }
 
-            await mediator.Send(new WriteChatCommand("/exit"));
+            await mediator.Send(new WriteChatCommand("/hideout"));
             return true;
         }
     }
